PMI587 - Refactor RP MVP script and restructure project to get a lean script while keeping the project structure simple
- Move RP MVP script into bin/ directory
- Refactor out piwik specific functions and group with existing PiwikClient
- Refactor out billing specific functions and group with rp_mapping (since its context is limited to billing)
- Refactor out how RP reports are generated into reports/rp
- Add a fixtures file in order to reuse test sample data across different test modules
