+++ performance/__init__.py
@@ -1,9 +1,3 @@
 from config import Config
-from .rp_federation_config import RPFederationConfig
-from .piwikclient import PiwikClient
 
 config = Config()
-
-piwik_client = PiwikClient(config)
-
-rp_mapping = RPFederationConfig(config)
